Refactor settings into dedicated package
