+++ multiaccounts/settings/settings_sync_handlers.go
@@ -1,4 +1,4 @@
-package accounts
+package settings
 
 import (
 	"errors"
