+++ protocol/messenger.go
@@ -35,6 +35,7 @@ import (
 	userimage "github.com/status-im/status-go/images"
 	"github.com/status-im/status-go/multiaccounts"
 	"github.com/status-im/status-go/multiaccounts/accounts"
+	"github.com/status-im/status-go/multiaccounts/settings"
 	"github.com/status-im/status-go/protocol/anonmetrics"
 	"github.com/status-im/status-go/protocol/audio"
 	"github.com/status-im/status-go/protocol/common"
@@ -973,7 +974,7 @@ func (m *Messenger) attachIdentityImagesToChatIdentity(context chatContext, ci *
 		return err
 	}
 
-	if s.ProfilePicturesShowTo == accounts.ProfilePicturesShowToNone {
+	if s.ProfilePicturesShowTo == settings.ProfilePicturesShowToNone {
 		m.logger.Info(fmt.Sprintf("settings.ProfilePicturesShowTo is set to '%d', skipping attaching IdentityImages", s.ProfilePicturesShowTo))
 		return nil
 	}
@@ -1019,7 +1020,7 @@ func (m *Messenger) attachIdentityImagesToChatIdentity(context chatContext, ci *
 		return fmt.Errorf("unknown ChatIdentity context '%s'", context)
 	}
 
-	if s.ProfilePicturesShowTo == accounts.ProfilePicturesShowToContactsOnly {
+	if s.ProfilePicturesShowTo == settings.ProfilePicturesShowToContactsOnly {
 		err := EncryptIdentityImagesWithContactPubKeys(ci.Images, m)
 		if err != nil {
 			return err
@@ -3425,7 +3426,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingCurrency)
 						logger.Debug("Handling SyncSettingCurrency", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.Currency, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.Currency, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingCurrency", zap.Error(err))
 							allMessagesProcessed = false
@@ -3441,7 +3442,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingGifFavorites)
 						logger.Debug("Handling SyncSettingGifFavorites", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.GifFavourites, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.GifFavourites, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingGifFavorites", zap.Error(err))
 							allMessagesProcessed = false
@@ -3457,7 +3458,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingGifRecents)
 						logger.Debug("Handling SyncSettingGifRecents", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.GifRecents, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.GifRecents, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingGifRecents", zap.Error(err))
 							allMessagesProcessed = false
@@ -3473,7 +3474,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingMessagesFromContactsOnly)
 						logger.Debug("Handling SyncSettingMessagesFromContactsOnly", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.MessagesFromContactsOnly, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.MessagesFromContactsOnly, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingMessagesFromContactsOnly", zap.Error(err))
 							allMessagesProcessed = false
@@ -3489,7 +3490,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingPreferredName)
 						logger.Debug("Handling SyncSettingPreferredName", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.PreferredName, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.PreferredName, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingPreferredName", zap.Error(err))
 							allMessagesProcessed = false
@@ -3505,7 +3506,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingPreviewPrivacy)
 						logger.Debug("Handling SyncSettingPreviewPrivacy", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.PreviewPrivacy, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.PreviewPrivacy, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingPreviewPrivacy", zap.Error(err))
 							allMessagesProcessed = false
@@ -3521,7 +3522,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingProfilePicturesShowTo)
 						logger.Debug("Handling SyncSettingProfilePicturesShowTo", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.ProfilePicturesShowTo, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.ProfilePicturesShowTo, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingProfilePicturesShowTo", zap.Error(err))
 							allMessagesProcessed = false
@@ -3537,7 +3538,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingProfilePicturesVisibility)
 						logger.Debug("Handling SyncSettingProfilePicturesVisibility", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.ProfilePicturesVisibility, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.ProfilePicturesVisibility, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingProfilePicturesVisibility", zap.Error(err))
 							allMessagesProcessed = false
@@ -3553,7 +3554,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingSendStatusUpdates)
 						logger.Debug("Handling SyncSettingSendStatusUpdates", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.SendStatusUpdates, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.SendStatusUpdates, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingSendStatusUpdates", zap.Error(err))
 							allMessagesProcessed = false
@@ -3569,7 +3570,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingStickerPacksInstalled)
 						logger.Debug("Handling SyncSettingStickerPacksInstalled", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.StickersPacksInstalled, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.StickersPacksInstalled, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingStickerPacksInstalled", zap.Error(err))
 							allMessagesProcessed = false
@@ -3585,7 +3586,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingStickerPacksPending)
 						logger.Debug("Handling SyncSettingStickerPacksPending", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.StickersPacksPending, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.StickersPacksPending, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingStickerPacksPending", zap.Error(err))
 							allMessagesProcessed = false
@@ -3601,7 +3602,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingStickersRecentStickers)
 						logger.Debug("Handling SyncSettingStickersRecentStickers", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.StickersRecentStickers, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.StickersRecentStickers, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingStickersRecentStickers", zap.Error(err))
 							allMessagesProcessed = false
@@ -3617,7 +3618,7 @@ func (m *Messenger) handleRetrievedMessages(chatWithMessages map[transport.Filte
 						setting := msg.ParsedMessage.Interface().(protobuf.SyncSettingTelemetryServerURL)
 						logger.Debug("Handling SyncSettingTelemetryServerURL", zap.Any("message", setting))
 
-						err := m.settings.SaveSyncSetting(accounts.TelemetryServerURL, setting.GetValue(), setting.GetClock())
+						err := m.settings.SaveSyncSetting(settings.TelemetryServerURL, setting.GetValue(), setting.GetClock())
 						if err != nil {
 							logger.Warn("failed to handle SyncSettingTelemetryServerURL", zap.Error(err))
 							allMessagesProcessed = false
@@ -5509,10 +5510,10 @@ func (m *Messenger) BloomFilter() []byte {
 	return m.transport.BloomFilter()
 }
 
-func (m *Messenger) getSettings() (accounts.Settings, error) {
+func (m *Messenger) getSettings() (settings.Settings, error) {
 	sDB, err := accounts.NewDB(m.database)
 	if err != nil {
-		return accounts.Settings{}, err
+		return settings.Settings{}, err
 	}
 	return sDB.GetSettings()
 }
