+++ services/web3provider/service.go
@@ -17,16 +17,21 @@ import (
 	"github.com/status-im/status-go/services/rpcfilters"
 )
 
-func NewService(appDB *sql.DB, rpcClient *rpc.Client, config *params.NodeConfig, accountsManager *account.GethManager, rpcFiltersSrvc *rpcfilters.Service, transactor *transactions.Transactor) *Service {
+func NewService(appDB *sql.DB, rpcClient *rpc.Client, config *params.NodeConfig, accountsManager *account.GethManager, rpcFiltersSrvc *rpcfilters.Service, transactor *transactions.Transactor) (*Service, error) {
+	accDB, err := accounts.NewDB(appDB)
+	if err != nil {
+		return nil, err
+	}
+
 	return &Service{
 		permissionsDB:   permissions.NewDB(appDB),
-		accountsDB:      accounts.NewDB(appDB),
+		accountsDB:      accDB,
 		rpcClient:       rpcClient,
 		rpcFiltersSrvc:  rpcFiltersSrvc,
 		config:          config,
 		accountsManager: accountsManager,
 		transactor:      transactor,
-	}
+	}, nil
 }
 
 type Service struct {
